[Feat/#37] 추천글 저장, 저장한 추천글 조회 API 연동

+++ app/src/main/java/com/example/resona/data/dto/PostResponseDto.kt
@@ -3,7 +3,7 @@ package com.example.resona.data.dto
 import com.google.gson.annotations.SerializedName
 
 data class PostResponseDto(
-    @SerializedName("postId") val postId: Int,
+    @SerializedName("postId") val postId: Long,
     @SerializedName("title") val title: String,
     @SerializedName("songTitle") val songTitle: String,
     @SerializedName("songUrl") val songUrl: String,
